f tester

+++ ft_print_pointer.c
@@ -1,6 +1,6 @@
 #include "ft_printf.h"
 
-int ft_print_pointer(unsigned long long ptr)
+int ft_print_pointer(unsigned long ptr)
 {
 	int i;
 
@@ -8,6 +8,6 @@ int ft_print_pointer(unsigned long long ptr)
 	if (ptr == 0)
 		return (write(1, "(nil)", 5));
 	i += write(1, "0x", 2);
-	i += ft_print_hex((unsigned long)ptr);
+	i += ft_print_hex(ptr);
 	return (i);
 }
